fix: fix array to string l.84 Evaluation

+++ logout.php
@@ -1,10 +1,11 @@
 <?php
-if($pizzaConfig){
-    $pizzaConfig = new DBconnection;
-    $alertMsg = $pizzaConfig->closeFct();
+if(isset($pizzaConfig)){
+    $pizzaConfig->closeFct();
 }
 session_start();
-unlink('image-qrcode.png');
+if (file_exists('image-qrcode.png')){
+    unlink('image-qrcode.png');
+}
 unset($_SESSION['name']);
 unset($_SESSION['restaurant']);
 unset($_SESSION['idSession']);
